Add support for dictionary application
Currently, we hijack the following syntax until we add parser support:

  e ((dict :: Class a))

Now any expression can be used as evidence. Consequently, a
`EvDictionary (HsExpr GhcTc)` constructor was added to the `EvTerm` data type
in `TcEvidence.hs`. This required an import of `HsExpr` from `HsExpr.hs`,
which caused an import cycle via `HsExtension.hs`. To break this cycle, the
`TcEvidence.hs-boot` module was added.