Infer the correct roles for dictionary data types
Type-class variables default to role nominal. If we use these roles for the
dictionary data type too, it means we cannot coerce dictionaries, which is
very useful.

Previously, we simply took the roles inferred for the type class and
downgraded each nominal to representational. This is incorrect, for example
when a type-class variable occurs in a Set in one of the type-class methods,
then its role should still be nominal instead of representational.

It turns out the roles were already being correctly inferred for the
dictionary type constructor, but we ignored it. Now we look up the roles
corresponding to the dictionary type constructor and pass them to the right
functions. Unfortunately, this requires change to the interface format, as the
roles are stored separately.